+++ CHANGELOG.md
@@ -6,5 +6,13 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## [Version 0.0.1] - 
 
+### Fixed:
+
+* (2024-03-01): fixed issue in make splits where the 2/3 image protocol
+  selection was backwards. Also improved protocol usage. Fixed issue where
+  after and negative images were never used.
+
+## [Version 0.0.1] - 
+
 ### Added
 * Initial version